// src/app/api/models/medical-specialty.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Detalles de la Medical Specialty a crear.
 */
export interface MedicalSpecialty {

  /**
   * Status of the specialty
   */
  active?: boolean;

  /**
   * Official Specialty Code
   */
  code: string;

  /**
   * Brief description of the specialty
   */
  description?: string;

  /**
   * Identificador único
   */
  id?: number;

  /**
   * Specialty Name
   */
  name: string;

  /**
   * Target Profession (DOCTOR, NURSE, BOTH)
   */
  targetProfession: string;
}